contracts/algorand/requirements.txt
py-algorand-sdk==1.8.0
pyteal==0.9.1
python-dotenv==0.19.0

# python -m venv venv
# source venv/bin/activate  # On Windows, use `venv\Scripts\activate`
# pip install -r requirements.txt
# python compile.py
# python deploy.py
